+++ src/unit_tests/mod.rs
@@ -1,8 +1,8 @@
 mod env;
 use env::*;
 
+mod catalog_with_filters;
 mod ctx;
-mod models;
 mod serde;
 
 mod link;
